[2538] (Feature) Added analytics call to server (#30)
* added analytics call

* added track validation

* updated invalid tracks

* feature: updated big decimal logic

* fixing build

* fixing tracks

* feature: updating tests

* feature: updated tests

* fixed public key logic

+++ core-methods/src/test/java/com/mercadopago/sdk/android/coremethods/domain/usecase/GetInstallmentsUseCaseTest.kt
@@ -18,7 +18,7 @@ internal class GetInstallmentsUseCaseTest {
     fun `invoke should call repository with correct parameters`() =
         runBlocking {
             val bin = "123456"
-            val amount = 1000L
+            val amount = 1000.0.toBigDecimal()
             val processingMode = "gateway"
 
             // Setup mock response
@@ -36,7 +36,7 @@ internal class GetInstallmentsUseCaseTest {
     fun `invoke should return error when repository fails`() =
         runBlocking {
             val bin = "123456"
-            val amount = 1000L
+            val amount = 1000.0.toBigDecimal()
             val processingMode = "gateway"
 
             // Setup mock to return an error
